de-dupe core and api-client as used by backend
