+++ apps/backend/src/events/projections/event-projector.ts
@@ -13,12 +13,13 @@ import { promises as fs } from 'fs';
 import * as path from 'path';
 import { didToAgent } from '../../utils/id-generator';
 import type { components } from '@semiont/api-client';
+import { compareAnnotationIds } from '@semiont/api-client';
 import type {
   DocumentEvent,
   StoredEvent,
   DocumentAnnotations,
 } from '@semiont/core';
-import { compareAnnotationIds, findBodyItem } from '@semiont/core';
+import { findBodyItem } from '@semiont/core';
 import type { ProjectionStorage, DocumentState } from '../../storage/projection-storage';
 
 type Document = components['schemas']['Document'];
